modified after code review - SuperHumanLab
